+++ src/TFGInfotecAbstractions/Interfaces/IAccountManager.cs
@@ -1,6 +1,6 @@
 ﻿namespace TFGInfotecAbstractions.Interfaces
 {
-	public interface IAccountManager
+    public interface IAccountManager
 	{
 		Task<Customer> CreateAccount(User user);
 		Task<string> SignIn(string email, string password);
