+++ src/TFGInfotecAbstractions/Models/AccountModels/BaseUser.cs
@@ -1,0 +1,11 @@
+﻿namespace TFGInfotecAbstractions.Models.AccountModels
+{
+    public class BaseUser : Entity
+    {
+        [Required]
+        public int UserId { get; set; }
+
+        [ForeignKey(nameof(UserId))]
+        public User? User { get; set; }
+    }
+}
