+++ src/TFGInfotecAbstractions/Models/AccountModels/Enums/Role.cs
@@ -1,0 +1,10 @@
+﻿namespace TFGInfotecAbstractions.Models.AccountModels.Enums
+{
+	public enum Role
+	{
+		[Display(Name = "Customer")]
+		Customer,
+		[Display(Name = "Admin")]
+		Admin
+	}
+}
