+++ src/TFGInfotecAbstractions/Models/AccountModels/Login.cs
@@ -1,0 +1,13 @@
+﻿namespace TFGInfotecAbstractions.Models.AccountModels
+{
+	public class Login
+    {
+        [Required]
+        [EmailAddress]
+        public string? Email { get; set; }
+
+        [Required]
+        [DataType(DataType.Password)]
+        public string? Password { get; set; }
+    }
+}
