+++ src/TFGInfotecAbstractions/Models/AccountModels/Registration.cs
@@ -1,0 +1,16 @@
+﻿namespace TFGInfotecAbstractions.Models.AccountModels
+{
+	public class Registration
+    {
+        [Required]
+        public string? Name { get; set; }
+
+        [Required]
+        [EmailAddress]
+        public string? Email { get; set; }
+
+        [Required]
+        [DataType(DataType.Password)]
+        public string? Password { get; set; }
+    }
+}
