--- src/TFGInfotecAbstractions/Models/BaseUser.cs
@@ -1,11 +1,0 @@
-﻿namespace TFGInfotecAbstractions.Models
-{
-	public class BaseUser : Entity
-	{
-		[Required]
-		public int UserId { get; set; }
-
-		[ForeignKey(nameof(UserId))]
-		public User? User { get; set; }
-	}
-}
