--- src/TFGInfotecAbstractions/Models/Customer.cs
@@ -1,6 +1,0 @@
-﻿namespace TFGInfotecAbstractions.Models
-{
-	public sealed class Customer : BaseUser
-	{
-	}
-}
