+++ src/TFGInfotecAbstractions/Models/DrinkModels/Customer.cs
@@ -1,0 +1,8 @@
+﻿using TFGInfotecAbstractions.Models.AccountModels;
+
+namespace TFGInfotecAbstractions.Models.DrinkModels
+{
+    public sealed class Customer : BaseUser
+    {
+    }
+}
