--- src/TFGInfotecAbstractions/Models/Image.cs
@@ -1,6 +1,0 @@
-﻿public sealed class Image : Entity
-{
-	[Required]
-	public required byte[] Data { get; set; }
-	public string? Type { get; set; }
-}
