--- src/TFGInfotecAbstractions/Models/Login.cs
@@ -1,15 +1,0 @@
-﻿using System.ComponentModel.DataAnnotations;
-
-namespace TFGInfotecAbstractions.Models
-{
-	public class Login
-	{
-		[Required]
-		[EmailAddress]
-		public string Email { get; set; }
-
-		[Required]
-		[DataType(DataType.Password)]
-		public string Password { get; set; }
-	}
-}
