--- src/TFGInfotecAbstractions/Models/Registration.cs
@@ -1,18 +1,0 @@
-﻿using System.ComponentModel.DataAnnotations;
-
-namespace TFGInfotecAbstractions.Models
-{
-	public class Registration
-	{
-		[Required]
-		public string Name { get; set; }
-
-		[Required]
-		[EmailAddress]
-		public string Email { get; set; }
-
-		[Required]
-		[DataType(DataType.Password)]
-		public string Password { get; set; }
-	}
-}
