+++ src/TFGInfotecAbstractions/Models/Usings.cs
@@ -1,0 +1,5 @@
+﻿global using TFGInfotecAbstractions.Models;
+global using TFGInfotecAbstractions.Models.AccountModels.Enums;
+global using TFGInfotecAbstractions.Models.DrinkModels;
+
+
