+++ src/TFGInfotecAbstractions/Usings.cs
@@ -2,5 +2,8 @@
 global using System.ComponentModel.DataAnnotations;
 global using System.ComponentModel.DataAnnotations.Schema;
 global using System.Diagnostics.CodeAnalysis;
-global using TFGInfotecAbstractions.Models;
+global using TFGInfotecAbstractions.Models.AccountModels;
+global using TFGInfotecAbstractions.Models.DishModels;
+global using System.Text.Json.Serialization;
+
