+++ src/TFGInfotecApi/Usings.cs
@@ -4,7 +4,9 @@ global using Microsoft.EntityFrameworkCore;
 global using System.Security.Claims;
 global using TFGInfotecAbstractions.Interfaces;
 global using TFGInfotecAbstractions.Interfaces.Services;
-global using TFGInfotecAbstractions.Models;
+global using TFGInfotecAbstractions.Models.AccountModels;
+global using TFGInfotecAbstractions.Models.DishModels;
+global using TFGInfotecAbstractions.Models.DrinkModels;
 global using TFGInfotecApi.Extensions;
 global using TFGInfotecCore.Managers;
 global using TFGInfotecCore.Services;
