+++ src/TFGInfotecInfrastructure/Context/ApplicationDbContext.cs
@@ -1,42 +1,43 @@
 ﻿namespace TFGInfotecInfrastructure.Context
 {
 	public class ApplicationDbContext : DbContext
-    {
+	{
 		public DbSet<User>? User { get; set; }
 		public DbSet<Drink>? Drinks { get; set; }
 		public DbSet<Dish>? Dishes { get; set; }
 		public DbSet<Account>? Account { get; set; }
 		public DbSet<Customer>? Customer { get; set; }
 
+		public DbSet<Image>? Image { get; set; }
+
 		public ApplicationDbContext(DbContextOptions<ApplicationDbContext> options)
-            : base(options)
-        {
-            try
-            {
-                var databaseCreator = Database.GetService<IDatabaseCreator>() as RelationalDatabaseCreator;
-                if (databaseCreator != null)
-                {
-                    if (!databaseCreator.CanConnect()) databaseCreator.Create();
-                    if (!databaseCreator.CanConnect()) databaseCreator.CreateTables();
-                }
-            }
-            catch (Exception ex)
-            {
-                Console.WriteLine(ex.Message);
-            }
-        }
+			: base(options)
+		{
+			try
+			{
+				var databaseCreator = Database.GetService<IDatabaseCreator>() as RelationalDatabaseCreator;
+				if (databaseCreator != null)
+				{
+					if (!databaseCreator.CanConnect()) databaseCreator.Create();
+					if (!databaseCreator.CanConnect()) databaseCreator.CreateTables();
+				}
+			}
+			catch (Exception ex)
+			{
+				Console.WriteLine(ex.Message);
+			}
+		}
 
 		protected override void OnModelCreating(ModelBuilder modelBuilder)
-        {
-            base.OnModelCreating(modelBuilder);
-
-			modelBuilder.Entity<Account>().ToTable("Account");
+		{
+			base.OnModelCreating(modelBuilder);
 
 			modelBuilder.Entity<User>()
 				.HasOne(u => u.Account)
 				.WithOne()
 				.HasForeignKey<User>(u => u.AccountId)
 				.OnDelete(DeleteBehavior.Cascade);
+
 			modelBuilder.Entity<Customer>()
 				.HasOne(b => b.User)
 				.WithOne()
@@ -49,6 +50,18 @@
 			modelBuilder.Entity<Account>()
 			.HasIndex(e => e.Email)
 			.IsUnique();
+
+			modelBuilder.Entity<Drink>()
+			.HasMany(r => r.Images)
+			.WithOne(i => i.Drink)
+			.HasForeignKey(i => i.ItemId)
+			.OnDelete(DeleteBehavior.Cascade);
+
+			modelBuilder.Entity<Dish>()
+			.HasMany(r => r.Images)
+			.WithOne(i => i.Dish)
+			.HasForeignKey(i => i.ItemId)
+			.OnDelete(DeleteBehavior.Cascade);
 		}
-    }
+	}
 }
